Encoding file type in mode
Fixed issue #38 by encoding the type within the `mode` property. This will allow the "stat-mode" package to function properly and therefore will let ftpd work correctly. This is a quick and dirty fix as the constants for File and Directory are hardcoded within this change - they are exported by "fs" but were empty when I tried to use them.